303-2, 304-2 날짜 가져오는 API 추가함. Parameter 추가함.

+++ src/main/java/kr/co/atg/apds/komipo_main/entity/pobject/P_getSpectrum.java
@@ -8,7 +8,7 @@ public class P_getSpectrum {
   Integer mptkey;
   String measdt;
   Integer spctype;
-  Integer fmax;
+  Double fmax;
   Integer windowfunc;
   Double deltafreq;
   Double deltaorder;
